requirements.txt
alembic==1.4.3
appdirs==1.4.3
async-lru==1.0.2
banal==1.0.1
CacheControl==0.12.6
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
dataset==1.4.1
distlib==0.3.0
distro==1.4.0
ed25519==1.5
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyaes==1.6.1
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
Pyrogram==1.0.7
PySocks==1.7.1
python-brightid==1.1.5
python-dateutil==2.8.1
python-editor==1.0.4
pytoml==0.1.21
requests==2.25.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.20
TgCrypto==1.2.2
toml==0.10.2
urllib3==1.26.2
webencodings==0.5.1
